Ac 586 task timeout implementation (#469)
* task timeout implemention & improved Tasks GetAllAsync

+++ src/WorkflowManager/Contracts/Models/TaskExecution.cs
@@ -72,6 +72,6 @@ namespace Monai.Deploy.WorkflowManager.Contracts.Models
         public DateTime Timeout { get => TaskStartTime.AddMinutes(TimeoutInterval); }
 
         [JsonProperty(PropertyName = "timeout_interval")]
-        public int TimeoutInterval { get; set; } = 0;
+        public double TimeoutInterval { get; set; } = 0;
     }
 }
